Switch from GraphRegressor to DirectForcefieldRegressor
