+++ orb_models/forcefield/forcefield_utils.py
@@ -1,8 +1,254 @@
-from typing import Optional, Tuple
-
+from typing import Dict, List, Literal, Mapping, Optional, Sequence, Tuple, Union
+from torch.nn import functional as F
 import torch
 
 from orb_models.forcefield import base, segment_ops
+from orb_models.forcefield.gns import MoleculeGNS
+
+
+
+def validate_regressor_inputs(
+    heads: Union[Sequence[torch.nn.Module], Mapping[str, torch.nn.Module]],
+    loss_weights: Dict[str, float],
+    ensure_grad_loss_weights: bool = False,
+):
+    """Validate the input heads and loss weights."""
+    if isinstance(heads, Sequence):
+        head_names = [head.target.fullname for head in heads]  # type: ignore
+    else:
+        head_names = list(heads.keys())
+        targets = [getattr(head, "target", None) for head in heads.values()]
+        if all(target is not None for target in targets):
+            target_names = [target.fullname for target in targets]  # type: ignore
+            if head_names != target_names:
+                raise ValueError(
+                    f"Head names and target names must match; got {head_names} and {target_names}"
+                )
+    if len(head_names) == 0:
+        raise ValueError("Model must have at least one output head.")
+    if len(head_names) != len(set(head_names)):
+        raise ValueError(f"Head names must be unique; got {head_names}")
+    unknown_keys = set(loss_weights.keys()) - set(head_names)  # type: ignore
+    if ensure_grad_loss_weights:
+        if (
+            "grad_forces" not in loss_weights.keys()
+            or "grad_stress" not in loss_weights.keys()
+        ):
+            raise ValueError("grad_forces and grad_stress must be in loss_weights .")
+        unknown_keys = unknown_keys - set(["grad_forces", "grad_stress"])
+    unknown_keys = unknown_keys - set(["rotational_grad"])  # not associated with a head
+    if unknown_keys:
+        raise ValueError(f"Loss weights for unknown targets: {unknown_keys}")
+
+
+def set_cutoff_layers(gns: MoleculeGNS, cutoff_layers: int):
+    """Set the number of message passing layers to keep."""
+    if cutoff_layers > gns.num_message_passing_steps:
+        raise ValueError(
+            f"cutoff_layers ({cutoff_layers}) must be less than or equal to"
+            f" the number of message passing steps ({gns.num_message_passing_steps})"
+        )
+    else:
+        gns.gnn_stacks = gns.gnn_stacks[:cutoff_layers]
+        gns.num_message_passing_steps = cutoff_layers  # type: ignore
+
+
+def split_prediction(pred: torch.Tensor, n_node: torch.Tensor):
+    """Split batched prediction back into per-system predictions."""
+    if len(pred) == len(n_node):
+        return torch.split(pred, 1, dim=0)
+    elif len(pred) == n_node.sum():
+        return torch.split(pred, n_node.cpu().tolist(), dim=0)
+    else:
+        raise ValueError(f"Unexpected length of prediction tensor: {len(pred)}")
+
+
+def mean_error(
+    pred: torch.Tensor,
+    target: torch.Tensor,
+    error_type: Literal["mae", "mse", "huber_0.01"],
+    batch_n_node: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    """Compute MAE or MSE for node or graph targets.
+
+    Args:
+        target: The target tensor.
+        pred: The prediction tensor.
+        batch_n_node: The number of nodes per graph. If provided, then a
+            nested aggregation is performed for node errors i.e. first we
+            average across nodes within each graph, then average across graphs.
+        error_type: The type of error to compute. Either "mae" or "mse" or "huber_x"
+            where x is the delta parameter for the huber loss.
+
+    Returns:
+        A scalar error for the whole batch.
+    """
+    if error_type.startswith("huber"):
+        huber_delta = float(error_type.split("_")[1])
+        error_type = "huber"  # type: ignore
+        assert huber_delta > 0.0, "HUBER_DELTA must be greater than 0.0"
+
+    error_function = {
+        "mae": lambda x, y: torch.abs(x - y),
+        "mse": lambda x, y: (x - y) ** 2,
+        "huber": lambda x, y: F.huber_loss(x, y, reduction="none", delta=huber_delta),
+    }[error_type]
+
+    errors = error_function(pred, target)
+    errors = errors.mean(dim=-1) if errors.dim() > 1 else errors
+
+    if batch_n_node is not None:
+        error = segment_ops.aggregate_nodes(
+            errors, batch_n_node, reduction="mean"
+        ).mean()
+    else:
+        error = errors.mean()
+
+    return error
+
+
+def bucketed_mean_error(
+    target: torch.Tensor,
+    pred: torch.Tensor,
+    bucket_by: Literal["target", "error"],
+    thresholds: List[float],
+    batch_n_node: Optional[torch.Tensor] = None,
+    error_type: Literal["mae", "mse"] = "mae",
+) -> Dict[str, torch.Tensor]:
+    """Compute MAE or MSE per-bucket, where each bucket is a range defined by thresholds.
+
+    The target can be a node-level or graph target. For node-level
+    targets, providing batch_n_node entails a nested-aggregation
+    of the error (first by node, then by graph).
+
+    Errors can be bucketed by their value, or by the value of the ground-truth target.
+    If bucketing by target, and the target is multi-dimensional, then the L2 norm
+    of the target is used to define the buckets.
+
+    Buckets are defined by a set of real-valued thresholds. For example,
+    bucket_by='error' and thresholds=[0.1, 10.0] creates 3 buckets:
+        - errors < 0.1
+        - 0.1 <= errors < 10.0
+        - errors >= 10.0
+
+    Args:
+        target: The target tensor.
+        pred: The prediction tensor.
+        bucket_by: The method for assigning buckets.
+        thresholds: The bucket edges. -inf and +inf are automatically added.
+        batch_n_node: The number of nodes per graph. If None, no nested aggregation is performed.
+        error_type: The type of error to compute. Either "mae" or "mse".
+
+    Returns:
+        A dictionary of metrics with entries of the form f"{error_type}_{bucket_name}", where
+        bucket name is a string representing the bucket edge values.
+    """
+    error_function = {
+        "mae": lambda x, y: torch.abs(x - y),
+        "mse": lambda x, y: (x - y) ** 2,
+    }[error_type]
+
+    errors = error_function(target, pred)
+
+    # If multi-dimensional, collapse down to a single error per graph/node
+    errors = errors.mean(dim=-1) if errors.dim() > 1 else errors
+
+    # Decide what to bucket by
+    if bucket_by == "target":
+        values_to_bucket_by = target.norm(dim=-1) if target.dim() > 1 else target
+    elif bucket_by == "error":
+        values_to_bucket_by = errors
+    else:
+        raise ValueError(f"Unknown bucket_by: {bucket_by}")
+
+    # Assign each element in the batch to a bucket index
+    bucket_edges = torch.tensor(
+        [-float("inf")] + thresholds + [float("inf")], device=errors.device
+    )
+    bucket_indices = torch.bucketize(values_to_bucket_by, bucket_edges, right=True) - 1
+
+    # Iterate over each bucket and compute its average error
+    metrics = {}
+    bucket_names = [
+        f"bucket_{p:.2f}-{q:.2f}" for p, q in zip(bucket_edges[:-1], bucket_edges[1:])
+    ]
+    for i, name in enumerate(bucket_names):
+        mask = bucket_indices == i
+        current_errors = errors[mask]
+
+        if batch_n_node is not None:
+            current_batch_n_node = segment_ops.aggregate_nodes(
+                mask.int(), batch_n_node, reduction="sum"
+            )
+            error = segment_ops.aggregate_nodes(
+                current_errors, current_batch_n_node, reduction="mean"
+            ).mean()
+        else:
+            error = current_errors.mean()
+
+        metrics[name] = error
+
+    return metrics
+
+
+def binary_accuracy(
+    pred: torch.Tensor, target: torch.Tensor, threshold: float = 0.5
+) -> float:
+    """Calculate binary accuracy between 2 tensors.
+
+    Args:
+        pred: the prediction tensor.
+        target: the tensor of target values.
+        threshold: Binary classification threshold. Default 0.5.
+
+    Returns:
+        mean accuracy.
+    """
+    return ((pred > threshold) == target).to(pred.dtype).mean().item()
+
+
+def categorical_accuracy(pred: torch.Tensor, target: torch.Tensor) -> float:
+    """Calculate accuracy for K class classification.
+
+    Args:
+        pred: the tensor of logits for K classes of shape (..., K)
+        target: tensor of integer target values of shape (...)
+
+    Returns:
+        mean accuracy.
+    """
+    pred_labels = torch.argmax(pred, dim=-1)
+    return (pred_labels == target.long()).to(pred.dtype).mean().item()
+
+
+def bce_loss(
+    pred: torch.Tensor, target: torch.Tensor, metric_prefix: str = ""
+) -> Tuple:
+    """Binary cross-entropy loss with accuracy metric."""
+    loss = torch.nn.BCEWithLogitsLoss()(pred, target.to(pred.dtype))
+    accuracy = binary_accuracy(pred, target)
+    return (
+        loss,
+        {
+            f"{metric_prefix}_accuracy": accuracy,
+            f"{metric_prefix}_loss": loss.item(),
+        },
+    )
+
+
+def cross_entropy_loss(
+    pred: torch.Tensor, target: torch.Tensor, metric_prefix: str = ""
+) -> Tuple:
+    """Cross-entropy loss with accuracy metric."""
+    loss = torch.nn.CrossEntropyLoss()(pred, target.long())
+    accuracy = categorical_accuracy(pred, target)
+    return (
+        loss,
+        {
+            f"{metric_prefix}_accuracy": accuracy,
+            f"{metric_prefix}_loss": loss.item(),
+        },
+    )
 
 
 def remove_fixed_atoms(
