+++ orb_models/forcefield/pretrained.py
@@ -1,34 +1,32 @@
 # flake8: noqa: E501
-from typing import Union, Optional
+from typing import Tuple, Union, Optional
 from functools import partial
 
 import torch
 from cached_path import cached_path
 
+from orb_models.forcefield.angular import SphericalHarmonics
+from orb_models.forcefield.atomic_system import SystemConfig
+from orb_models.forcefield.conservative_regressor import ConservativeForcefieldRegressor
+from orb_models.forcefield.direct_regressor import DirectForcefieldRegressor
 from orb_models.forcefield.featurization_utilities import (
     get_device,
     gaussian_basis_function,
 )
 from orb_models.forcefield.gns import MoleculeGNS
-from orb_models.forcefield.graph_regressor import (
-    GraphHead,
-    GraphRegressor,
-)
-from orb_models.forcefield.forcefield_heads import EnergyHead, ForceHead
+from orb_models.forcefield.forcefield_heads import ConfidenceHead, EnergyHead, ForceHead, GraphHead, StressHead
+from orb_models.forcefield.rbf import BesselBasis
 from orb_models.utils import set_torch_precision
 
 
 def load_model_for_inference(
-    model: GraphRegressor,
+    model: Union[DirectForcefieldRegressor, ConservativeForcefieldRegressor],
     weights_path: str,
     device: Union[torch.device, str, None] = None,
     precision: str = "float32-high",
     compile: Optional[bool] = None,
-) -> GraphRegressor:
+) -> Union[DirectForcefieldRegressor, ConservativeForcefieldRegressor]:
     """Load a pretrained model from a local path or a wandb artifact.
-
-    The model class is inferred from the hydra config. This config is expected
-    to contain a `model` key that, when instantiated, is a valid `base.ModelMixin`.
 
     Args:
         model: The model to load the weights into.
@@ -77,26 +75,26 @@ def load_model_for_inference(
 def orb_v2_architecture(
     num_message_passing_steps: int = 15,
     device: Optional[Union[torch.device, str]] = None,
-) -> GraphRegressor:
+) -> DirectForcefieldRegressor:
     """Orb-v2 architecture."""
-    model = GraphRegressor(
-        heads=[
-            EnergyHead(
+    model = DirectForcefieldRegressor(
+        heads={
+            "energy": EnergyHead(
                 latent_dim=256,
                 num_mlp_layers=1,
                 mlp_hidden_dim=256,
                 predict_atom_avg=True,
                 activation="ssp",
             ),
-            ForceHead(
-                latent_dim=256,
+            "forces": ForceHead(
+                latent_dim=256, 
                 num_mlp_layers=1,
                 mlp_hidden_dim=256,
                 remove_mean=True,
                 remove_torque_for_nonpbc_systems=True,
                 activation="ssp",
             ),
-            GraphHead(
+            "stress": GraphHead(
                 latent_dim=256,
                 num_mlp_layers=1,
                 mlp_hidden_dim=256,
@@ -104,7 +102,7 @@ def orb_v2_architecture(
                 node_aggregation="mean",
                 activation="ssp",
             ),
-        ],
+        },
         model=MoleculeGNS(
             num_node_in_features=256,
             num_node_out_features=3,
@@ -121,6 +119,7 @@ def orb_v2_architecture(
             activation="ssp",
             mlp_norm="layer_norm",
         ),
+        pair_repulsion=False,
     )
     device = get_device(device)
     if device is not None and device != torch.device("cpu"):
@@ -131,19 +130,290 @@ def orb_v2_architecture(
     return model
 
 
+def orb_v3_conservative_architecture(
+    latent_dim: int = 256,
+    base_mlp_hidden_dim: int = 1024,
+    base_mlp_depth: int = 2,
+    head_mlp_hidden_dim: int = 256,
+    head_mlp_depth: int = 1,
+    num_message_passing_steps: int = 5,
+    activation: str = "silu",
+    device: Optional[torch.device] = None,
+) -> ConservativeForcefieldRegressor:
+    """The orb-v3 conservative architecture."""
+    model = ConservativeForcefieldRegressor(
+        heads={
+            "energy": EnergyHead(
+                latent_dim=latent_dim,
+                num_mlp_layers=head_mlp_depth,
+                mlp_hidden_dim=head_mlp_hidden_dim,
+                predict_atom_avg=True,
+                activation=activation,
+            ),
+            "confidence": ConfidenceHead(
+                latent_dim=latent_dim,
+                num_mlp_layers=head_mlp_depth,
+                mlp_hidden_dim=head_mlp_hidden_dim,
+                activation=activation,
+            ),
+        },
+        model=MoleculeGNS(
+            latent_dim=latent_dim,
+            num_message_passing_steps=num_message_passing_steps,
+            num_mlp_layers=base_mlp_depth,
+            mlp_hidden_dim=base_mlp_hidden_dim,
+            rbf_transform=BesselBasis(
+                r_max=6.0,
+                num_bases=8,
+            ),
+            angular_transform=SphericalHarmonics(
+                lmax=3,
+                normalize=True,
+                normalization="component",
+            ),
+            outer_product_with_cutoff=True,
+            use_embedding=True,
+            interaction_params={
+                "distance_cutoff": True,
+                "attention_gate": "sigmoid",
+            },
+            node_feature_names=["feat"],
+            edge_feature_names=["feat"],
+            activation=activation,
+            mlp_norm="rms_norm",
+        ),
+        ensure_grad_loss_weights=False,
+        pair_repulsion=True,
+    )
+    if device is not None and device != torch.device("cpu"):
+        model.cuda(device)
+    else:
+        model = model.cpu()
+
+    return model
+
+
+def orb_v3_direct_architecture(
+    latent_dim: int = 256,
+    base_mlp_hidden_dim: int = 1024,
+    base_mlp_depth: int = 2,
+    head_mlp_hidden_dim: int = 256,
+    head_mlp_depth: int = 1,
+    num_message_passing_steps: int = 5,
+    activation: str = "silu",
+    device: Optional[torch.device] = None,
+) -> DirectForcefieldRegressor:
+    """The orb-v3 architecture, defaulting to a direct model."""
+    model = DirectForcefieldRegressor(
+        heads={
+            "energy": EnergyHead(
+                latent_dim=latent_dim,
+                num_mlp_layers=head_mlp_depth,
+                mlp_hidden_dim=head_mlp_hidden_dim,
+                predict_atom_avg=True,
+                activation=activation,
+            ),
+            "forces": ForceHead(
+                latent_dim=latent_dim,
+                num_mlp_layers=head_mlp_depth,
+                mlp_hidden_dim=head_mlp_hidden_dim,
+                remove_mean=True,
+                remove_torque_for_nonpbc_systems=True,
+                activation=activation,
+            ),
+            "stress": StressHead(
+                latent_dim=latent_dim,
+                num_mlp_layers=head_mlp_depth,
+                mlp_hidden_dim=head_mlp_hidden_dim,
+                node_aggregation="mean",
+                activation=activation,
+            ),
+            "confidence": ConfidenceHead(
+                latent_dim=latent_dim,
+                num_mlp_layers=head_mlp_depth,
+                mlp_hidden_dim=head_mlp_hidden_dim,
+                activation=activation,
+            ),
+        },
+        model=MoleculeGNS(
+            latent_dim=latent_dim,
+            num_message_passing_steps=num_message_passing_steps,
+            num_mlp_layers=base_mlp_depth,
+            mlp_hidden_dim=base_mlp_hidden_dim,
+            rbf_transform=BesselBasis(
+                r_max=6.0,
+                num_bases=8,
+            ),
+            angular_transform=SphericalHarmonics(
+                lmax=3,
+                normalize=True,
+                normalization="component",
+            ),
+            outer_product_with_cutoff=True,
+            use_embedding=True,
+            interaction_params={
+                "distance_cutoff": True,
+                "attention_gate": "sigmoid",
+            },
+            node_feature_names=["feat"],
+            edge_feature_names=["feat"],
+            activation=activation,
+            mlp_norm="rms_norm",
+        ),
+        pair_repulsion=True,
+    )
+    if device is not None and device != torch.device("cpu"):
+        model.cuda(device)
+    else:
+        model = model.cpu()
+
+    return model
+
+def orb_v3_conservative_20_omat(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[ConservativeForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Conservative 20 max neighbors OMAT."""
+    model = orb_v3_conservative_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
+
+def orb_v3_conservative_inf_omat(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[ConservativeForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Conservative with effectively unlimited neighbors, trained on OMAT.
+
+    'Effectively unlimited' means that the model will use all neighbors within 6A
+    the cutoff radius. Empirically, for the training distribution, 120 is sufficient.
+    """
+    model = orb_v3_conservative_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=120)
+
+def orb_v3_direct_20_omat(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Direct 20 max neighbors OMAT."""
+    model = orb_v3_direct_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
+
+def orb_v3_direct_inf_omat(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Direct with effectively unlimited neighbors, trained on OMAT.
+
+    'Effectively unlimited' means that the model will use all neighbors within 6A
+    the cutoff radius. Empirically, for the training distribution, 120 is sufficient.
+    """
+    model = orb_v3_direct_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=120)
+
+def orb_v3_conservative_20_mpa(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[ConservativeForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Conservative 20 max neighbors MPTraj + Alexandria."""
+    model = orb_v3_conservative_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
+
+def orb_v3_conservative_inf_mpa(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[ConservativeForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Conservative with effectively unlimited neighbors, trained on MPTraj + Alexandria.
+
+    'Effectively unlimited' means that the model will use all neighbors within 6A
+    the cutoff radius. Empirically, for the training distribution, 120 is sufficient.
+    """
+    model = orb_v3_conservative_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=120)
+
+def orb_v3_direct_20_mpa(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Direct 20 max neighbors MPTraj + Alexandria."""
+    model = orb_v3_direct_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
+
+def orb_v3_direct_inf_mpa(
+    weights_path: str = "",  # noqa: E501
+    device: Union[torch.device, str, None] = None,
+    precision: str = "float32-high",
+    compile: Optional[bool] = None,
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB v3 Direct with effectively unlimited neighbors, trained on MPTraj + Alexandria.
+
+    'Effectively unlimited' means that the model will use all neighbors within 6A
+    the cutoff radius. Empirically, for the training distribution, 120 is sufficient.
+    """
+    model = orb_v3_direct_architecture(device=device)
+    model = load_model_for_inference(
+        model, weights_path, device, precision=precision, compile=compile
+    )
+
+    return model, SystemConfig(radius=6.0, max_num_neighbors=120)
+
+
+
 def orb_v2(
     weights_path: str = "https://orbitalmaterials-public-models.s3.us-west-1.amazonaws.com/forcefields/orb-v2-20241011.ckpt",  # noqa: E501
     device: Union[torch.device, str, None] = None,
     precision: str = "float32-high",
     compile: Optional[bool] = None,
-) -> GraphRegressor:
-    """Load ORB v2."""
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB v2 Direct with 20 max neighbors, trained on MPTraj + Alexandria."""
     model = orb_v2_architecture(device=device)
     model = load_model_for_inference(
         model, weights_path, device, precision=precision, compile=compile
     )
+    # TODO (BEN): update all functions to return SystemConfig
+    # TODO (BEN): search repo for max_num_neighbors and avoid any hardcoding
 
-    return model
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
 
 
 def orb_mptraj_only_v2(
@@ -151,8 +421,8 @@ def orb_mptraj_only_v2(
     device: Union[torch.device, str, None] = None,
     precision: str = "float32-high",
     compile: Optional[bool] = None,
-):
-    """Load ORB MPTraj Only v2."""
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB MPTraj Only v2 Direct with 20 max neighbors, trained on MPTraj."""
 
     return orb_v2(weights_path, device, precision=precision, compile=compile)
 
@@ -162,14 +432,14 @@ def orb_d3_v2(
     device: Union[torch.device, str, None] = None,
     precision: str = "float32-high",
     compile: Optional[bool] = None,
-) -> GraphRegressor:
-    """Load ORB D3 v2."""
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB D3 v2 Direct with 20 max neighbors, trained on MPTraj + Alexandria."""
     model = orb_v2_architecture(device=device)
     model = load_model_for_inference(
         model, weights_path, device, precision=precision, compile=compile
     )
 
-    return model
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
 
 
 def orb_d3_sm_v2(
@@ -177,14 +447,14 @@ def orb_d3_sm_v2(
     device: Union[torch.device, str, None] = None,
     precision: str = "float32-high",
     compile: Optional[bool] = None,
-) -> GraphRegressor:
-    """Load ORB D3 v2."""
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB D3 small v2 with 20 max neighbors, trained on MPTraj + Alexandria."""
     model = orb_v2_architecture(num_message_passing_steps=10, device=device)
     model = load_model_for_inference(
         model, weights_path, device, precision=precision, compile=compile
     )
 
-    return model
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
 
 
 def orb_d3_xs_v2(
@@ -192,14 +462,14 @@ def orb_d3_xs_v2(
     device: Union[torch.device, str, None] = None,
     precision: str = "float32-high",
     compile: Optional[bool] = None,
-) -> GraphRegressor:
-    """Load ORB D3 xs v2."""
+) -> Tuple[DirectForcefieldRegressor, SystemConfig]:
+    """Load ORB D3 xs v2 with 20 max neighbors, trained on MPTraj + Alexandria."""
     model = orb_v2_architecture(num_message_passing_steps=5, device=device)
     model = load_model_for_inference(
         model, weights_path, device, precision=precision, compile=compile
     )
 
-    return model
+    return model, SystemConfig(radius=6.0, max_num_neighbors=20)
 
 
 def _deprecated_model(model_name: str):
@@ -255,6 +525,17 @@ def orb_v1_mptraj_only(
 
 
 ORB_PRETRAINED_MODELS = {
+    # most performant orb-v3 omat models
+    "orb-v3-conservative-20-omat": orb_v3_conservative_20_omat, 
+    "orb-v3-conservative-inf-omat": orb_v3_conservative_inf_omat,
+    "orb-v3-direct-20-omat": orb_v3_direct_20_omat,
+    "orb-v3-direct-inf-omat": orb_v3_direct_inf_omat,
+    # orb-v3 mptraj + alexandria models
+    "orb-v3-conservative-20-mpa": orb_v3_conservative_20_mpa,
+    "orb-v3-conservative-inf-mpa": orb_v3_conservative_inf_mpa,
+    "orb-v3-direct-20-mpa": orb_v3_direct_20_mpa,
+    "orb-v3-direct-inf-mpa": orb_v3_direct_inf_mpa,
+    # less performant orb-v2 models
     "orb-v2": orb_v2,
     "orb-d3-v2": orb_d3_v2,
     "orb-d3-sm-v2": orb_d3_sm_v2,
